+++ libs/gmana/lib/helpers/helpers.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:math';
 
 part 'id.dart';
-part 'reg.dart';
-part 'types.dart';
-part 'validators.dart';
 part 'is.dart';
+part 'reg.dart';
+part 'typedef.dart';
+part 'validators.dart';