+++ libs/gmana/lib/helpers/types.dart
@@ -1,3 +1,0 @@
-part of 'helpers.dart';
-
-typedef GMap = Map<String, dynamic>;